Make timeout waiting for sync a non-error exit
Once a commit has resulted from a job, an operation has effectively
succeeded. Various `fluxctl ..` commands wait for the commit to be
applied, however, which can take arbitrarily long, and may time
out. When used in scripts, this can mean the script fails when the
operation really succeeded.

So: when the client times out waiting for sync, print some advice to
stderr and exit(0) instead of returning an error.