Reconcile git cloning part of tests
Lots of tests use the git mirroring, and have to wait for it to report
that it's ready before proceeding. We see this flake all the time, and
it seemed odd that it couldn't complete a git clone on the local
filesystem within five seconds.

The helper function was trying every 100ms, five times, but reporting
that it had tried for five seconds if it failed. Instead, try every
100ms, thirty times, and report accordingly.