Implement the new FHR format for cluster resources
We still have largely parallel models of things that are resources in
the cluster, and things that are defined in the git repo. Both models
need to be able to interpret FluxHelmRelease values in terms of
containers.

Because it's the same type (map[string]interface{}) being interpreted
in both places, we can write one procedure to do the interpretation,
and just use the interpretation in different ways. So: factor out a
visitor-style procedure, and use it for cluster and repo resources.