Put containers in FluxHelmReleases in stable order
Release verification expects the containers in the `after` state to be
the same as in the `before` state. Since we are interpreting
containers given a map, and Go makes maps unordered (or rather, it
enumerates the keys in a non-deterministic order), we need to sort the
keys to make a stable result.