Add `flux release --interactive`
This adds an `--interactive` option to the `release` command that lets
you pick and choose which containers to update. It uses the new
`containers` specification in the /v9/update-manifest call.

Once the containers are selected, the release is sent off with
the constraint that the state of the containers need to be the exact
same as when the results were presented to the user. If anything changed
in the meanwhile, the whole release will not be processed.