Read default image creds for each request
Part of the point of being able to provide a docker config is to be
able to mount it from a volume, and thereby be able to update it on
the fly (e.g., with a sidecar).

But this won't work if it's read only at startup. So, just check it's
OK when we start, and otherwise read it every time we are asked for
credentials.