Remove chart index in favour of that in gh-pages
We use the gh-pages branch as a chart repository, by generating and
checking in tarballs and merging the new packages into the index
(chart/index.yaml).

Since the charts live in master, but the packages in gh-pages, there's
a bit of faffery to update the latter. One solution is to keep the
index here, do the packaging, then switch to gh-pages and add the
tarball and updated index.

But it's easier to maintain things from within gh-pages branch. I've
added a script there which makes the tarball and updates the index --
which means we don't need the index here.