Provide read-only modes of using git
For the helm operator (and in some places in fluxd) we need only
read-only access to the git repo. Indeed in some deployments, the helm
operator will be given a read-only SSH key.

So: give the git mirroring (git.Repo) a read-only option, which means
it will

 - consider itself ready when it's been cloned, without checking
   whether it can write to the upstream repo;
 - not let you make a working clone (because it's assumed you won't be
   able to push commits from it)

We still need a way to get a copy of the repo, so the helm operator
can look at the files. For that purpose, add a bare-bones type
`git.Export`, with which all you can do is get the directory it's been
cloned to.