Provide a synchronous way to ready a git mirror
For fluxd, we have a git mirroring implementation that keeps track of
its ready state. This is because when fluxd starts, it may not have a
viable SSH key, so it needs to be able to retry in the background,
while it does everything else that it can.

In the helm operator, and in tests, it's more useful to be able to
clone the repo and succeed or fail synchronously. This should remove
some flakes from the tests, which previously had to start the repo's
loop, and check back with it to see if it was ready -- which would
occasionally fail to be the case within the timeout.