Fluxctl should search for pod labels added by helm chart
The Flux Helm chart does not have a name label on it, but we still want the flux instance
to be discoverable by fluxctl.

This change makes it so that if fluxctl does not find a pod by the labels
`name in (flux,fluxd,weave-flux-agent)`, it searches for `app=flux`.

If it does not find either label, it returns a nice message:

```
$ ./fluxctl --k8s-fwd-ns lol list-controllers
Error: Flux pod not found for labels in namespace lol:
  name in (flux,fluxd,weave-flux-agent)
  app=flux
Run 'fluxctl list-controllers --help' for usage.
$
```

Fixes #1242