Improve image with zero timestamp safeguard
Since we'd expect an image with a zero timestamp to be shuffled to the
end of the list, we probably won't hit the safeguard just put in,
which checks whether the latest image has a zero timestamp. Instead,
while we're looking for the current image so we can log the
comparison, also check if _any_ images have a zero timestamp, and skip
that container if they do.

Ideally we'd avoid having data with this problem in the first place,
but let's have defence in depth.