Fix git.Export(sha1)
`git clone -b ...` does not accept a revision -- it must be a branch
or tag -- so we can't use that to export a repo at a revision, as we'd
like to do for the helm operator's sync loop.

Instead we'll have to clone without an argument, and checkout the
desired revision.