Use exactly one clone of the git repo for syncing
If we just clone the repo at the branch HEAD whenever the operator
wants to apply a change, we may end up undoing, or redoing, work,
since the ChartChangeSync only advances the revision it looks at when
it's ready to.

To avoid that, pass the desired reconcilations to the ChartChangeSync
and have it do the work, at the revision it's currently working from.

The mechanism is to send the resource in question on an unbuffered
channel, and receive it in the Run loop of ChartChangeSync. This has a
couple of flaws:

 - the event handlers have to wait for the Run loop to get around to
   doing the work (or time out and retry);

 - there's not much point in running more than one worker to process
   events, since they all get serialised through the ChartChangeSync
   anyway (not quite true -- deletes are done straight away because
   they don't need to consult the git repo)

But it is good enough for now.