Return all containers found in FluxHelmRelease
When intepreting the `values` part of a FluxHelmRelease, _all_
structures than can be interpreted as specifying a container
(something mentioning an image, essentially) should be returned, since
that is both more useful and more in line with expectations.

Prior to this commit, the implementation would return _either_ a
container defined at the top level (i.e., an `image` field), _or_ a
set of containers defined in other fields. But this is not adequate
even for our own flux chart.