Check if boolean policies equal true
It should be possible to set the annotation, e.g.,

    flux.weave.works/automated: 'false'

and expect automation to be off for that resource. However, since we
were checking only for the presence of the automation policy, and not
its value, any value was interpreted as on.

I've changed the name of the method used, so that it's a break from
the previous use (and to flush out where it was used).