Record unusable image manifests, rather than fail
At present the registry client considers a non-Linux/AMD64 image an
error, which provokes the image "cache" warmer to fail for that entire
image repo. As more and more image repositories have Windows images,
this mean flux won't be able to automate or even report on many
workloads.

What we actually want to do is just exclude unusable images from the
result; and for bonus points, record the fact that it wasn't a usable
image. To this end, I've wrapped the cache entries in a struct that
also includes an "excluded reason". If that field is present, the
fetch is considered a success, but the image is not included in the
result for the repo.

The cache entry version doesn't need to be bumped, because old entries
can be parsed as new entries. Happily, no-one has to refresh their
image DB from scratch.