Get imagePullSecrets from the service account too
It is possible, and often more economical, to give ImagePullSecrets to
a service account (perhaps the default for the namespace), rather than
each workload individually.

So that flux will still find these secrets, look up the service
account, and its imagePullSecrets, if any. This involves one more RPC
to the API server (to get the service account), per workload.