Keep track of image pull secrets already seen
Each time we refresh the list of images to fetch metadata for, we do this:

 * create a map of image->credentials
 * for each workload,
   * look up all the imagePullSecrets associated with it and merge
     them together into credentials
   * for each image used in the workload, merge with the entry in the
     map for that image

This works, but it uses far more API calls than it really needs to,
because secrets are very likely to be reused as imagePullSecrets for
different workloads.

To avoid doing extra lookups, keep track of secrets already seen
within each namespace, and don't ask for them again when they are
mentioned afterwards.