Run install when the release is (likely) missing
Before I rewrote things to drive all the syncing through chartsync,
syncing used `release.Exists(...)` to determine whether to Install or
Upgrade a given release. That worked by coincidence since any error,
including but not only a "not found" error, was treated as a negative
result.

However: it was a useful coincidence, because _just_ checking the
error return value means a "not found" will exit early rather than go
on to install the (possibly) missing release. So, ignore the error
value, and choose whether to install or upgrade based on whether there
was a release value returned or not, as `Exists` did.