Allow multiple config paths within a git repo
The idea is to let people overlay bits of configuration by supplying
more than one `--git-path` argument.

Much of the machinery for dealing with manifest files (and
directories) already expected more than one path; but, it wasn't
necessarily convenient to supply all of them in one go, so I've
updated the Manifests interface and followed that through.

The git path appears in API results -- it's just an opaque string,
though, so joining the now multiple paths and returning that should be
fine.