Get individual namespaces when given whitelist
Asking for the full list of namespaces requires a cluster-scoped
permission of listing namespaces; however, a common scenario for using
the whitelist is that you want to restrict permissions.

If we simply Get the whitelisted namespaces, ignoring those we're
forbidden to see (or that don't exist, as before), we don't need the
cluster-scoped permission and can just be given permissions per
namespace.

The trade is that we do an API request per whitelisted namespace. I
expect there to be relatively few, though, so I don't think this is a
huge deal.