Correctly store Excluded manifests
I noticed that fluxd was continually fetching image manifests that
were "excluded", that is, had the wrong arch or otherwise weren't
suitable.

The design in #1265 was to mark those manifests as excluded, but count
them as successful fetches; and, to save the fact in memcached. The
implementation failed on three counts:

 - it would unconditionally use a cached result, even if excluded;

 - it was using the ID from excluded results (which have no ID value)
   as a key, so they got effectively thrown away anyway;

 - the encoding of results to JSON omitted the ExcludedReason field,
   because they embed the image.Info struct and thus inherit its JSON
   encoding methods.

This commit corrects those problems.