Log a missing namespace only when first noticed
If we log a warning every time a whitelisted is missing, there may be
an awful lot of repeated warnings. Instead, keep track of which
namespaces have been seen to be missing (resetting when they are seen
again), and log only when the namespace was not known to be missing.