Put some trace logging into the cache refresh code
To see whether the image cache refresh is doing what you expect, it's
useful for it to log what it's doing as it does it (and why it does
things). This commit adds some trace messages, switched on with the
`--registry-trace` flag, and tidies up some of the other logging.