Include initContainers in images to fetch
Symptom: images used in initContainers, but not elsewhere, don't get
metadata fetched for them, and don't get automated updates.

Fix: the ImagesToFetch method of the cluster, supplied to the image DB
as the way to get a list of all the images to fetch, didn't look in
initContainers. So, make it do that.