Add ListServicesWithOptions in API v11
Adds the endpoint `ListServicesWithOptions` supporting `namespace` and
`services` as a parameter. `services` filters the result by resource
ids.

It introduces a new API v11 which builds on top of v10 but adds the
`ListServicesWithOptions` call. It is polyfilled back to API v6 by
using `ListServices` and then manually removing the unwanted services.