Use git repo as specified in each FHR
This is a change in design: rather than having a git repo that you
provide to the operator via command-line args, let people specify for
each FluxHelmRelease if it comes from a git repo, and if so, where in
the repo the chart lives (branch and path).

To adapt to this design, instead of having a single *git.Export which
is updated when the single (command-line argument) git repo has new
commits, we use the newly minted git.Mirrors to dynamically keep track
of all git repos in play.

Within that scheme there are a few choices to be made:

 * How do you provide keys for authenticating with git+ssh?

I have punted on this for now, and left it to the user to mount a key
into the correct location (same as before). This means you can use at
most one identity, and if that's a "deploy key" for github, at most
one repo, since github doesn't let you reuse keys.

You can of course mount more keys and a ssh_config which points at
them, but it could be made easier.

 * Do we keep a mirror per FluxHelmRelease?

That would be safe; however, it would most likely do much more work
than necessary, since it's expected that some git repos will contain
more than one chart.

Instead, one mirror is kept per git repository. If each FHR came with
its own SSH key, we might need to segregate instances of the same git
repo (since some might have valid creds and some not). But this is
good enough for now.

 * Can we pool clones?

Each FHR can refer to a distinct path and branch, even if it shares
the git repo. So to be safe, I have given each FHR its own clone;
these are updated when the files in the branch and path have changed
since the last update.