Set conditions on FluxHelmReleases
Conditions are a way for a controller to report the status of
particular resource. In the case of FluxHelmReleases, there's (at
least) a couple of things we want to report on:

 - has the controller managed to get the chart mentioned in the FHR?
 - has the controller managed to release the chart?

These can be used to determine whether the FluxHelmRelease is "ready",
for various definitions of "ready" (e.g., has it released the chart
version given in the FHR? If not, why not?)

To be able to set conditions reporting the status of charts from Helm
repos, the bit doing the condition setting must know what happens when
attempting to fetch the charts. That bit is the `chartsync` package,
which has the core loop reconciling the state of the cluster, the git
mirroring, and (now) the cached charts. So, move the chart fetching
code, and its call sites, to `chartsync`.

Moving the chart fetching code from `release` to `chartsync` makes
sense for another reason: it means the argument to
`release.Install(..)` can be a path to the chart, uniformly, and no
interpretation of the chart source in FluxHelmRelease is required
there.

When calculating the changes to conditions, we have to take into
account those already in effect. As an approximation to checking the
cluster value each time, mutate the local value, so that later changes
will be considered against those earlier.