Add repositories.yaml to helm-operator image
We want to be able to update the dependencies of a Helm chart. Helm
(whether the client library or the command-line tool) makes it
difficult to do this without also having other equipment, notably the
repositories.yaml file which records aliases and possibly credentials
for Helm repos (e.g., "stable").

This commit adds a default repositories.yaml file to the image,
containing just the conventional location of "stable".