Move garbage collection into cluster implementation
It's a nice separation to have the algorithm for syncing independent
of the cluster implementation; however, when tracking things to delete
it introduces some awkwardness.

Firstly, it means the Cluster interface has to have a method for
fetching all the candidates for garbage collection. This was
`ExportByLabel`. Secondly, the method `cluster.Sync` was designed
around deciding what to apply and what to delete, then supplying all
the operations in one call; whereas the garbage-collecting algorithm
must first apply changes, then calculate what's left over and delete
it. This isn't wrong as such, but it does expose a mismatch between
Sync interface and its use.

While keeping the algorithm as similar as possible, this commit moves
the garbage collection phase into the cluster/kubernetes
implementation of `Sync(...)`. This means `Cluster.Sync` and its
argument `SyncDef` can be simplified since much of the bookkeeping
will be done inside the implemention.

This commit also wires a flag through from the command line to the
cluster implementation, so one can switch garbage collection on. The
wiring bypasses the sync package, for simplicity -- it will always
calculate checksums.

Some included tidies:

 - move the ambiguously situated and named type apiObject to
where it's used, renames it, and narrows its scope to just the file
it's in.
 - use map-valued mixins rather than assembling text
 - turn the temporary Printfs into logger.Log invocations,
   and remove messages that are just reporting business as usual.