Support both new and old custom resources
Instead of requiring people to cut across from the old version of the
helm operator to the new, support both in fluxd.

This means you can not only have different versions of the custom
resources, you can keep running the old operator, while you transition
to the new.

To keep the `kind` part of Resource IDs unambiguous, I have renamed
the new custom resource to `HelmRelease` (it was already in the API
`flux.weave.works/v1beta`). This is the same as the kind bitnami's
helm operator uses, but the `apiVersion` is different (so Kubernetes
won't confuse them), and it's unlikely you'd be running both
operators.