Implement and document per-operator authentication
As a stop-gap in the absence of per-_resource_ authentication, this
commit makes it possible to provide repository credentials by mounting
a `repositories.yaml` file into the operator container.

I've replaced the aspirational documentation on using
`chartPullSecret`s with practical documentation on how to mount that
aforementioned file and SSH keys; and, updated the example deployment
with example volumes and mounts.