Bring tests up to date with sync changes
To start, this changes types/definitions to be in line with the
changes to `cluster.Cluster.Sync()` and `cluster.SyncDef`.

After that, to verify that the cluster implementation does in fact
apply and delete resources appropriately, we must have some fairly
elaborate scaffolding. The cluster must be able to ask for the
resources extant, as well as have apply and delete _stick_ (i.e, be
observable in subsequent queries).

To do that, we have to create fake clients, and make sure they are
primed to respond in the expected way. This amounts to

 - making sure the core client can enumerate the namespaces used in
   the test; and,

 - giving the dynamic client the appropriate API resources, so it will
   know to ask for the objects used in the test.

It's also necessary to have an up-to-date implementation of the fake
dynamic client; prior versions had a bug that made these tests
impossible. To avoid having a cascading dependency on the very latest
kubernetes code, I've simply taken one file we need, at the version
that works (see cluster/kubernetes/fakedynamicclient_test.go).

The pay-off is that having set up this scaffolding, we can test
syncing quite directly: the test itself boils down to "sync these fake
manifests, then see whether they got applied in the cluster".