Bump to kubeyaml 0.5.1 to fix (F)HR updating
kubeyaml 0.5.1 has a fix for updating images that include ports in
(Flux)HelmRelease manifests. So: bump that in the Dockerfile and shim
(`/bin/kubeyaml`), and add a test that it does account for such images
when updating.

Side issue in update_test.go: calling `t.Parallel()` _before_ invoking
the test case meant that the closed-over loop variable could be
reassigned before the test case was run.