Bump baked-in kubectl to 1.11.3
Now that Kubernetes 1.12 is GA, we are running into problems where
YAMLs that are valid for 1.12 are rejected by kubectl. For example, it
doesn't like validation of custom resources -- something we use
ourselves in the HelmRelease CRD.

kubectl versions are officially compatible with one version behind and
one version ahead; this effectively drops support for Kubernetes 1.9,
though I would expect it to still work. (For reference, Kubernetes 1.10
was released in March 2018.)