Clear up use of annotations and labels for sync
Although it would be cluster-neutral to use policies, the checksum and
stack have to be implemented as distinct bits of metadata, since
policies get encoded as annotations and we want to select (via the
Kubernetes API) by labels. Therefore, deal with the annotation and
label used, explicitly.

This _does_ mean we don't have to create `resource.Resource`s when we
want to examine what is in the cluster, though -- we can just pass
around the things that we got from the Kubernetes API, with a shim for
getting the metadata we care about.