Avoid deleting resources that did not apply OK
It's possible that when we go to apply a particular manifest, there's
some reason it doesn't work. In that case, we don't want to delete it,
since it's clearly intended to still exist.

In other words, anything that was represented by a manifest is kept,
whether it was applied successfully or not. I still have a case
checking if the checksum matches, so it can be reported in a log
message.