Add explicit flag for TLS insecure skip verify
We accept a list of image registry hosts for which to use HTTP; in
some cases, people also want to switch off TLS host certificate
verification, e.g., they have a setup in which the image manifests are
on a host using TLS but with a self-signed cert.

This extra flag allows people to switch TLS host cert verification off
(it's a bad idea to do that by default) explicitly, when the registry
is listed as an insecure host.