Respect "ignore" when syncing and GCing
Prior to chopping and changing how Sync works, and adding garbage
collection, the "ignore" annotation was implemented by exporting the
cluster resources (which only yielded namespaces and workloads!) and
checking for each manifest in the repo whether it, or its namesake in
the cluster, had the "ignore" annotation.

I have reinstated this logic in the Sync method. If the cluster
resource or repo manifest is marked with "ignore", the manifest will
not be staged for application; and, if a cluster resource is marked
with "ignore", it will not be deleted during GC.

To check the cluster resources, we have to have _all_ of them, and not
just those that were synced by flux in the past, thus
`getResourcesBySelector` (which will get all resources, when given an
empty selector). When it comes to GC, we only care about things that
were previously synced by flux, so using `getResourcesInStack` is
fine.