Use checksum per resource
In using a checksum for a whole stack (which at present is everything
in the repo), it makes the syncing very sensitive to changes. Since it
would mostly be just changing annotations, I don't think this is a
problem for correctness, but it will keep Kubernetes busy, updating
annotations every time anything changes.

It's not expensive to calculate the checksum for each resource, as it
goes past. We don't actually use it for change detection -- it's only
consulted when it comes to garbage collect items, to check if it was
updated earlier (and if not, a warning is issued).