Get AWS auth token for each ECR registry
Instead of asking for the AWS region and set of registry IDs (really
account IDs), take these from the registry URLs as they come up.

This means we have to make a token request for each region; so, keep
track of the expiry, per region. And so we don't keep asking when
there's a failure, embargo requests for a region (for ten minutes)
whenever we fail to get a token.

The arguments `--aws-region` and `--aws-registry-id` are now ignored,
since the values are taken from the registry URLs; but, the plumbing
for those values is left in, because the arguments are useful with an
altered meaning of "only these regions/accountIDs". This change will
be made in a subsequent commit.