Restrict ECR scans based on --aws-* args
After the last commit, which takes the region and account ID from the
registry URL, this commit repurposes the arguments --aws-region and
(renamed) --aws-account-id to be "allow lists".

They differ in the default policy: if no account IDs are given, all
accounts IDs are allowed; whereas, if no region is given, the allowed
regions is taken to be the "cluster region".

What is the cluster region? It's the region as detected from
configuration (so it can be overridden), or as given by the EC2
metadata service, as a last resort. The rationale for always
restricting the region is that using images from a different region
will incurs costs, so it's better to encourage people to be deliberate
about it.

 - [ ] does it still work with minikube (or does it fail at creating wrapper)