Remove Manifests#ParseManifests
Since we're no longer relying on Export when syncing (which was always
a little circuitous), there's no need for ParseManifests.

Where it was used for testing, it's OK to use the Kubernetes-specific
procedure instead (cluster/kubernetes/resource#ParseMultidoc), or
avoid the Export->ParseManifests sequence altogether and treat things
as opaque strings/bytes.