Account for namespace defaulting
It's possible to have manifests for namespaces resources in your git
repo, but not specify a namespace in them. In this case, `kubectl`,
and thereby fluxd, will transparently assign a default namespace
(either from kubeconfig, or if not given there, the default fallback
of `"default"`). When we're syncing, we need to be able to know which
namespace these resources will end up in, so we can identify them
during garbage collection.

The basic scheme is: in a few strategic places, let a fallback
namespace be supplied, which is used for anything that doesn't have
one. To avoid having to check which resources are supposed to have
namespaces (e.g., namespaces themselves), use the fallback to identify
everything that doesn't have one, either from a file or from the
cluster.

(The alternative is to figure out precisely which things need to have
namespaces, e.g., using the discovery API, and give only those a
default namespace. So long as the namespace is used _only_ to identify
resources, giving everything a namespace is equivalent. And since lots
of existing code assumes there's always a namespace, I leant that
way.)

Since the fallback namespace is decided by `kubectl`, that's what we
use to discover it.