Warn users about external changes in sync tag
Multiple users have reported an unexpectedly high number of changes in the git
respository's sync tag.

This has been attributed to multiple fluxd instances, running in separate
clusters, using the same git repository and sync tag.

In addition, multiple fluxd instances using the same tag is a bad idea in
general, since it will lead to commits being missed by fluxd.

This change tracks the value of the sync tag and prints a warning when detecting
an external change.