Assign namespaces after parsing
So we can separate parsing from processing at least minimally, this
commit makes assigning a namespace to resources (that need one) a
post-processing step, in `Manifests.LoadManifests`. It also makes it a
more reliable operation than before, since it looks up whether or not
a particular resource is _supposed_ to have a namespace, according to
the Kubernetes API, before assigning the configured default.