Move the shutdown wait to the end of helm-op main
The general scheme in main.go is to start a bunch of goroutines,
giving them a shutdown channel; then wait until there's a reason to
exit, close the shutdown channel, and wait until the goroutines exit.

This wasn't working, because the last stanza in main.go was _not_ run
in a goroutine. This commit fixes that, and moves the graceful
shutdown bit to the end, for clarity.