Record the Revision in the status
At current state, we're unable to track what exact revision of code is
running on our clusters.

This should not be a problem, if we'd rely on chart "version" being the
source of truth, however sadly it is possible to make changes to the
chart without bumping a version. Possibly should be bumped with a CI/CD
tool, but it also has a potential to fail.

Our use case:

```
Merge to master happens
CI picks the changes up
CI runs unit tests
CI marks merge as staging ready by promoting a tag
Flux running on staging cluster/namespace picks up the changes
CI (that is not exposed to the internet - webhooks are a no no) is
   assured the change has been applied
CI runs the integration tests
CI marks merge as staging ready by promoting a tag
Flux running on production cluster/namespace picks up the changes
CI notifies of completion
```

This could be done better. Instead of us passing around a string from
function to function, we could convert the current code obtaining the
revision into it's own function and call it at the very end to gather
this again. Open to interpretation.

Signed-off-by: Rafal Proszowski <paroxp@gmail.com>