Take ignore policy into account while fetching
Before this change having a HelmRelease annotated with both 'automated'
and 'ignore' put automated releases to a halt for all resources known
to Flux.

The reason for this was the image check failing while comparing with a
non-existent controller, resulting in an early error return; which
short circuited image updates for everything else.