Use git repo config to identify SyncSet
Using a constant for identifying sync sets means that if more than one
fluxd is running in a cluster, they will garbage collect each other's
resources.

We want to be able to delimit which resources are _this_ fluxd's
responsibility. Using the git config, including the branch and paths,
means fluxd will only garbage collect files that are in the bit of the
repo it is applying. That assumes no other fluxd is looking at
(exactly) the same directories in the same git URL -- which is a fair
assumption, since that would obviously be a bad config and cause
problems, even without considering garbage collection.