Bump Helm API package and binary to v2.13.0
There was a breaking change in 2.13 (see
https://github.com/weaveworks/flux/issues/1774 and the linky therein).

It appears that the Helm operator _as it is_ will work OK with the
newer Tiller (including TLS). But it's about time to make the leap to
a newer version, since 2.9.1 is just about a year old.

I have also verified that a helm operator built with this commit works
in a cluster with Tiller 2.9.1, i.e., with existing installations.