Introduce `valuesFrom` to HelmRelease CRD spec
`valuesFrom` supports different types of sources: `configMapKeyRef`,
`secretKeyRef` and `externalSourceRef`. Where the latter can be used to
retrieve a values file from an URL by setting the
`externalSourceRef.URL`. The array of sources is merged in specified
order, with the `.spec.values` getting merged last.

For backwards compatiblity the old `valuesFromSecrets` are transformed
into `secretKeyRef`s and lined up in front of any specified
`valuesFrom` sources.

On all source types the `<sourceRef>.Optional` can be set to `true` to
allow the release of a Helm chart to continue if the values file could
not be found.