Use FQRN to prevent errors when branch = label
If the label and branch were set to the same name, Flux would not
create the sync tag as the revision lookup for `HEAD` and `<tag>`
returned the same revision.

This is due to how git looks up symbolic ref names, giving priority
to heads over tags. Using the (almost) fully qualified ref name will
tell git to only look for matches of that type.

https://git-scm.com/docs/gitrevisions#Documentation/gitrevisions.txt-emltrefnamegtemegemmasterememheadsmasterememrefsheadsmasterem