Ensure release has not changed before upgrade run
It is possibile a release vanishes or is modified between the start
of a release and the actual upgrade itself, for example due to the
`HelmRelease` getting removed by fluxd's garbage collector.

Check if the release is still there and has not been modified in the
time we prepared for the upgrade.