Separate ECR auto-auth from ECR include/exclude
It is useful to be able to constrain the image repos scanned, even
when you don't need or want to use the AWS authentication.

This commit makes the constraints operate independently of AWS
authentication, by using a "pre-flight check" to determine whether the
AWS auth should be used, but applying the constraints either way. The
preflight check is also used to exit from main() if
`--registry-require=ecr` is set and the AWS API is not available.