Detect changes made to git source in HelmRelease
Before this change, changes made to the git chart source in a
`HelmRelease` did go unnoticed. Now, in case the URL or ref in the
`HelmRelease` changes, the clone is removed so a new one can be made
that matches the updated URL and ref. As a bonus, in case a
`HelmRelease` is removed, the clone will now be removed too.