Verify scanned keys in same build step as scan
ssh-keyscan can get keys which then fail verification (e.g., if it
misses out a host for some reason), but nonetheless cache the
intermediate image, which will cause subsequent builds to fail. This
is a pain to recover from.

This commit changes the Dockerfiles such the verification is done in
the same step as the keyscan -- so if it fails, the intermediate image
won't be cached, and subsequent builds will do the keyscan again.