Do not trigger update event on status changes
Before this change, and due to the fact that we no longer compare
the spec of the given HelmRelease against the spec of the old one,
every status and condition update would enqueue a new release job.

By updating the subresource instead of the resource, we no longer
trigger the update event, causing no unwanted trigger of a release.