Refresh memcache server(s) periodically
The memcache client resolves hostnames on creation, so if DNS is not
available when fluxd is starting, memcache operations will fail
consistently.

This was less of an issue when we (mostly) used the service discovery
(SRV record) lookup, since that resets the server list once every
minute or so. But when we started recommending the "fixed" server
lookup (a.k.a., hostname resolution) it became more prevalent.

The simplest fix, which I've done here, is to refresh the hostname
lookup in the same way as the service discovery implementation.